remove duplicate heap_4_lock file, and do not invoke the malloc failed handler when the allocation size is 0
